+++ CHANGES.rst
@@ -13,6 +13,7 @@ development
 - [ui] Improve "Share this search" dialog
 - [ui] Disable dummy exception catcher introduced accidentally when migration to jQuery3
 - [ui] Fix rendering in quick access mode, e.g. ``/view/pn/EP0666666``
+- [ui] Improve usability of "Share this search" and "Fetch result numbers" in the context of search results
 
 
 2018-09-28 0.152.0